@@ -7,7 +7,7 @@ using Microsoft.Build.Logging.StructuredLogger;
 
 namespace BinlogTool
 {
-    public class SaveFiles
+    public class SaveFiles : BinlogToolCommandBase
     {
         private string[] args;
 
@@ -18,7 +18,8 @@ namespace BinlogTool
 
         public void Run(string binlog, string outputDirectory, bool reconstruct = false)
         {
-            if (string.IsNullOrEmpty(binlog) || !File.Exists(binlog))
+            Build build = this.ReadBuild(binlog, false);
+            if (build == null)
             {
                 return;
             }
@@ -29,9 +30,6 @@ namespace BinlogTool
                 Directory.CreateDirectory(outputDirectory);
             }
 
-            binlog = Path.GetFullPath(binlog);
-
-            var build = BinaryLog.ReadBuild(binlog);
             SaveFilesFrom(build, outputDirectory);
 
             if (reconstruct)
@@ -181,7 +179,13 @@ namespace BinlogTool
 
         private void SaveFilesFrom(Build build, string outputDirectory)
         {
-            foreach (var file in build.SourceFiles.Values.OrderBy(f => f.FullPath))
+            var files = build.SourceFiles;
+            if (files == null)
+            {
+                return;
+            }
+
+            foreach (var file in files.OrderBy(f => f.FullPath))
             {
                 var filePath = file.FullPath;
                 if (filePath.EndsWith(".metaproj", StringComparison.OrdinalIgnoreCase) ||
