@@ -20,6 +20,7 @@ using Microsoft.Build.Logging.StructuredLogger;
 using Microsoft.Language.Xml;
 using Mono.Cecil;
 using StructuredLogViewer.Core.ProjectGraph;
+using TPLTask = System.Threading.Tasks.Task;
 
 namespace StructuredLogViewer.Controls
 {
@@ -29,22 +30,27 @@ namespace StructuredLogViewer.Controls
         public TreeViewItem SelectedTreeViewItem { get; private set; }
         public string LogFilePath => Build?.LogFilePath;
 
-        private ScrollViewer scrollViewer;
-
         private SourceFileResolver sourceFileResolver;
         private ArchiveFileResolver archiveFile => sourceFileResolver.ArchiveFile;
         private PreprocessedFileManager preprocessedFileManager;
         private NavigationHelper navigationHelper;
 
+        private MenuItem searchMenuGroup;
+        private MenuItem copyMenuGroup;
+        private MenuItem gotoMenuGroup;
         private MenuItem copyItem;
         private MenuItem copySubtreeItem;
         private MenuItem viewSubtreeTextItem;
         private MenuItem searchInSubtreeItem;
+        private MenuItem searchInNodeByNameItem;
+        private MenuItem searchThisNode;
         private MenuItem excludeSubtreeFromSearchItem;
+        private MenuItem excludeNodeByNameFromSearch;
         private MenuItem goToTimeLineItem;
         private MenuItem goToTracingItem;
         private MenuItem copyChildrenItem;
         private MenuItem sortChildrenItem;
+        private MenuItem filterChildrenItem;
         private MenuItem copyNameItem;
         private MenuItem copyValueItem;
         private MenuItem viewSourceItem;
@@ -52,10 +58,16 @@ namespace StructuredLogViewer.Controls
         private MenuItem openFileItem;
         private MenuItem copyFilePathItem;
         private MenuItem preprocessItem;
+        private MenuItem searchNuGetItem;
         private MenuItem runItem;
         private MenuItem debugItem;
         private MenuItem hideItem;
         private MenuItem showTimeItem;
+        private MenuItem favoriteItem;
+        private MenuItem unfavoriteItem;
+        private MenuItem favoriteSharedItem;
+        private MenuItem unfavoriteSharedItem;
+
         private ContextMenu sharedTreeContextMenu;
         private ContextMenu filesTreeContextMenu;
 
@@ -71,14 +83,22 @@ namespace StructuredLogViewer.Controls
 
             searchLogControl.ExecuteSearch = (searchText, maxResults, cancellationToken) =>
             {
+                if (Build.SearchIndex is { } index)
+                {
+                    index.MaxResults = maxResults;
+                    index.MarkResultsInTree = SettingsService.MarkResultsInTree;
+                    var indexResults = index.FindNodes(searchText, cancellationToken);
+                    PrecalculationDuration = index.PrecalculationDuration;
+                    return indexResults;
+                }
+
                 var search = new Search(
                     new[] { Build },
                     Build.StringTable.Instances,
                     maxResults,
-                    SettingsService.MarkResultsInTree
-                    //, Build.StringTable // disable validation in production
-                    );
+                    SettingsService.MarkResultsInTree);
                 var results = search.FindNodes(searchText, cancellationToken);
+                PrecalculationDuration = search.PrecalculationDuration;
                 return results;
             };
             searchLogControl.ResultsTreeBuilder = BuildResultTree;
@@ -120,10 +140,10 @@ namespace StructuredLogViewer.Controls
             DataContext = build;
             Build = build;
 
-            if (build.SourceFilesArchive != null)
+            // first try to see if the source archive was embedded in the log
+            if (build.SourceFiles != null)
             {
-                // first try to see if the source archive was embedded in the log
-                sourceFileResolver = new SourceFileResolver(build.SourceFiles.Values);
+                sourceFileResolver = new SourceFileResolver(build.SourceFiles);
             }
             else
             {
@@ -138,21 +158,34 @@ namespace StructuredLogViewer.Controls
 
             // Search Log | Properties and Items | Find in Files
             sharedTreeContextMenu = new ContextMenu();
+            sharedTreeContextMenu.Opened += SharedTreeContextMenu_Opened;
+            favoriteSharedItem = new MenuItem { Header = "Add to Favorites" };
+            unfavoriteSharedItem = new MenuItem { Header = "Remove from Favorites" };
+            var sharedCopyItem = new MenuItem() { Header = "Copy" };
             var sharedCopyAllItem = new MenuItem() { Header = "Copy All" };
             var sharedCopySubtreeItem = new MenuItem() { Header = "Copy subtree" };
+            favoriteSharedItem.Click += (s, a) => AddToFavorites();
+            unfavoriteSharedItem.Click += (s, a) => RemoveFromFavorites();
+            sharedCopyItem.Click += (s, a) => Copy();
             sharedCopyAllItem.Click += (s, a) => CopyAll();
             sharedCopySubtreeItem.Click += (s, a) => CopySubtree();
+            sharedTreeContextMenu.AddItem(favoriteSharedItem);
+            sharedTreeContextMenu.AddItem(unfavoriteSharedItem);
+            sharedTreeContextMenu.AddItem(sharedCopyItem);
             sharedTreeContextMenu.AddItem(sharedCopyAllItem);
             sharedTreeContextMenu.AddItem(sharedCopySubtreeItem);
 
             // Files
             filesTreeContextMenu = new ContextMenu();
+            var filesCopyItem = new MenuItem { Header = "Copy" };
             var filesCopyAllItem = new MenuItem { Header = "Copy All" };
             var filesCopyPathsItem = new MenuItem { Header = "Copy file paths" };
             var filesCopySubtreeItem = new MenuItem { Header = "Copy subtree" };
+            filesCopyItem.Click += (s, a) => Copy();
             filesCopyAllItem.Click += (s, a) => CopyAll();
             filesCopyPathsItem.Click += (s, a) => CopyPaths();
             filesCopySubtreeItem.Click += (s, a) => CopySubtree();
+            filesTreeContextMenu.AddItem(filesCopyItem);
             filesTreeContextMenu.AddItem(filesCopyAllItem);
             filesTreeContextMenu.AddItem(filesCopyPathsItem);
             filesTreeContextMenu.AddItem(filesCopySubtreeItem);
@@ -160,85 +193,119 @@ namespace StructuredLogViewer.Controls
             // Build Log
             var contextMenu = new ContextMenu();
             contextMenu.Opened += ContextMenu_Opened;
+            searchMenuGroup = new() { Header = "Search" };
+            copyMenuGroup = new() { Header = "Copy" };
+            gotoMenuGroup = new() { Header = "Go to" };
+
             copyItem = new MenuItem() { Header = "Copy" };
             copySubtreeItem = new MenuItem() { Header = "Copy subtree" };
             viewSubtreeTextItem = new MenuItem() { Header = "View subtree text" };
             searchInSubtreeItem = new MenuItem() { Header = "Search in subtree" };
             excludeSubtreeFromSearchItem = new MenuItem() { Header = "Exclude subtree from search" };
-            goToTimeLineItem = new MenuItem() { Header = "Go to timeline" };
-            goToTracingItem = new MenuItem() { Header = "Go to tracing" };
+            excludeNodeByNameFromSearch = new MenuItem() { Header = "Exclude node from search" };
+            searchInNodeByNameItem = new MenuItem() { Header = "Search in this node." };
+            searchThisNode = new MenuItem() { Header = "Search This Node" };
+            goToTimeLineItem = new MenuItem() { Header = "Timeline" };
+            goToTracingItem = new MenuItem() { Header = "Tracing" };
             copyChildrenItem = new MenuItem() { Header = "Copy children" };
             sortChildrenItem = new MenuItem() { Header = "Sort children" };
+            filterChildrenItem = new MenuItem() { Header = "Filter children (Ctrl+F)" };
             copyNameItem = new MenuItem() { Header = "Copy name" };
             copyValueItem = new MenuItem() { Header = "Copy value" };
             viewSourceItem = new MenuItem() { Header = "View source" };
             viewFullTextItem = new MenuItem { Header = "View full text" };
             showTimeItem = new MenuItem() { Header = "Show time and duration" };
+            favoriteItem = new MenuItem() { Header = "Add to Favorites" };
+            unfavoriteItem = new MenuItem() { Header = "Remove from Favorites" };
             openFileItem = new MenuItem() { Header = "Open File" };
             copyFilePathItem = new MenuItem() { Header = "Copy file path" };
             preprocessItem = new MenuItem() { Header = "Preprocess" };
+            var nugetImage = new System.Windows.Shapes.Path
+            {
+                Data = (Geometry)Application.Current.FindResource("NuGetGeometry"),
+                Stroke = (Brush)Application.Current.FindResource("NuGet"),
+                Fill = (Brush)Application.Current.FindResource("NuGet"),
+                Width = 16,
+                Height = 16,
+                StrokeThickness = 1
+            };
+            searchNuGetItem = new MenuItem() { Header = "Search project.assets.json", Icon = nugetImage };
             hideItem = new MenuItem() { Header = "Hide" };
             runItem = new MenuItem() { Header = "Run" };
             debugItem = new MenuItem() { Header = "Debug" };
             copyItem.Click += (s, a) => Copy();
-            copySubtreeItem.Click += (s, a) => CopySubtree(treeView);
+            copySubtreeItem.Click += (s, a) => CopySubtree(ActiveTreeView);
             viewSubtreeTextItem.Click += (s, a) => ViewSubtreeText();
             searchInSubtreeItem.Click += (s, a) => SearchInSubtree();
             excludeSubtreeFromSearchItem.Click += (s, a) => ExcludeSubtreeFromSearch();
+            excludeNodeByNameFromSearch.Click += (s, a) => ExcludeNodeByNameFromSearch();
+            searchInNodeByNameItem.Click += (s, a) => SearchInNodeByName();
+            searchThisNode.Click += (s, a) => SearchThisNode();
             goToTimeLineItem.Click += (s, a) => GoToTimeLine();
             goToTracingItem.Click += (s, a) => GoToTracing();
             copyChildrenItem.Click += (s, a) => CopyChildren();
             sortChildrenItem.Click += (s, a) => SortChildren();
+            filterChildrenItem.Click += (s, a) => FilterChildren();
             copyNameItem.Click += (s, a) => CopyName();
             copyValueItem.Click += (s, a) => CopyValue();
             viewSourceItem.Click += (s, a) => Invoke(treeView.SelectedItem as BaseNode);
             viewFullTextItem.Click += (s, a) => ViewFullText(treeView.SelectedItem as BaseNode);
             showTimeItem.Click += (s, a) => ShowTimeAndDuration();
+            favoriteItem.Click += (s, a) => AddToFavorites();
+            unfavoriteItem.Click += (s, a) => RemoveFromFavorites();
             openFileItem.Click += (s, a) => OpenFile();
             copyFilePathItem.Click += (s, a) => CopyFilePath();
             preprocessItem.Click += (s, a) => Preprocess(treeView.SelectedItem as IPreprocessable);
+            searchNuGetItem.Click += (s, a) => SearchNuGet(treeView.SelectedItem as IProjectOrEvaluation);
             runItem.Click += (s, a) => Run(treeView.SelectedItem as Task, debug: false);
             debugItem.Click += (s, a) => Run(treeView.SelectedItem as Task, debug: true);
             hideItem.Click += (s, a) => Delete();
 
+            contextMenu.AddItem(favoriteItem);
+            contextMenu.AddItem(unfavoriteItem);
             contextMenu.AddItem(runItem);
             contextMenu.AddItem(debugItem);
             contextMenu.AddItem(viewSourceItem);
             contextMenu.AddItem(viewFullTextItem);
             contextMenu.AddItem(openFileItem);
-            contextMenu.AddItem(preprocessItem);
-            contextMenu.AddItem(searchInSubtreeItem);
-            contextMenu.AddItem(excludeSubtreeFromSearchItem);
-            contextMenu.AddItem(goToTimeLineItem);
-            contextMenu.AddItem(goToTracingItem);
-            contextMenu.AddItem(copyItem);
-            contextMenu.AddItem(copySubtreeItem);
-            contextMenu.AddItem(copyFilePathItem);
-            contextMenu.AddItem(viewSubtreeTextItem);
-            contextMenu.AddItem(copyChildrenItem);
+            gotoMenuGroup.AddItem(preprocessItem);
+            contextMenu.AddItem(searchMenuGroup);
+            searchMenuGroup.AddItem(searchNuGetItem);
+            searchMenuGroup.AddItem(searchInSubtreeItem);
+            searchMenuGroup.AddItem(searchInNodeByNameItem);
+            searchMenuGroup.AddItem(searchThisNode);
+            searchMenuGroup.AddItem(excludeSubtreeFromSearchItem);
+            searchMenuGroup.AddItem(excludeNodeByNameFromSearch);
+            contextMenu.AddItem(gotoMenuGroup);
+            gotoMenuGroup.AddItem(goToTimeLineItem);
+            gotoMenuGroup.AddItem(goToTracingItem);
+            contextMenu.AddItem(copyMenuGroup);
+            copyMenuGroup.AddItem(copyItem);
+            copyMenuGroup.AddItem(copySubtreeItem);
+            copyMenuGroup.AddItem(copyFilePathItem);
+            gotoMenuGroup.AddItem(viewSubtreeTextItem);
+            copyMenuGroup.AddItem(copyChildrenItem);
             contextMenu.AddItem(sortChildrenItem);
-            contextMenu.AddItem(copyNameItem);
-            contextMenu.AddItem(copyValueItem);
-            contextMenu.AddItem(showTimeItem);
+            contextMenu.AddItem(filterChildrenItem);
+            copyMenuGroup.AddItem(copyNameItem);
+            copyMenuGroup.AddItem(copyValueItem);
+            gotoMenuGroup.AddItem(showTimeItem);
             contextMenu.AddItem(hideItem);
 
-            var existingTreeViewItemStyle = (Style)Application.Current.Resources[typeof(TreeViewItem)];
-            var treeViewItemStyle = new Style(typeof(TreeViewItem), existingTreeViewItemStyle);
-            treeViewItemStyle.Setters.Add(new Setter(TreeViewItem.IsExpandedProperty, new Binding("IsExpanded") { Mode = BindingMode.TwoWay }));
-            treeViewItemStyle.Setters.Add(new Setter(TreeViewItem.IsSelectedProperty, new Binding("IsSelected") { Mode = BindingMode.TwoWay }));
-            treeViewItemStyle.Setters.Add(new Setter(TreeViewItem.VisibilityProperty, new Binding("IsVisible") { Mode = BindingMode.TwoWay, Converter = new BooleanToVisibilityConverter() }));
+            var treeViewItemStyle = TreeViewExtensions.CreateTreeViewItemStyleWithEvents<BaseNode, TreeViewItem>();
 
             treeViewItemStyle.Setters.Add(new EventSetter(MouseDoubleClickEvent, (MouseButtonEventHandler)OnItemDoubleClick));
-            treeViewItemStyle.Setters.Add(new EventSetter(PreviewMouseRightButtonDownEvent, (MouseButtonEventHandler)OnPreviewMouseRightButtonDown));
-            treeViewItemStyle.Setters.Add(new EventSetter(RequestBringIntoViewEvent, (RequestBringIntoViewEventHandler)TreeViewItem_RequestBringIntoView));
             treeViewItemStyle.Setters.Add(new EventSetter(KeyDownEvent, (KeyEventHandler)OnItemKeyDown));
 
             treeView.ContextMenu = contextMenu;
             treeView.ItemContainerStyle = treeViewItemStyle;
-            treeView.KeyDown += TreeView_KeyDown;
+            treeView.KeyUp += TreeView_KeyDown;
             treeView.SelectedItemChanged += TreeView_SelectedItemChanged;
             treeView.GotFocus += TreeView_GetFocus;
             treeView.AddHandler(TreeViewItem.SelectedEvent, (RoutedEventHandler)TreeViewItem_Selected);
+
+            findTextBox.KeyDown += FindTextBox_KeyDown;
+            searchLogControl.searchTextBox.KeyUp += SearchTextBox_KeyUp;
 
             ActiveTreeView = treeView;
 
@@ -277,13 +344,20 @@ on the node will navigate to the corresponding source code associated with the n
 
 More functionality is available from the right-click context menu for each node.
 Right-clicking a project node may show the 'Preprocess' option if the version of MSBuild was at least 15.3.";
-                build.Unseal();
 #if DEBUG
                 text = build.StringTable.Intern(text);
 #endif
-                build.AddChild(new Note { Text = text });
-                build.Seal();
+                var folder = new Folder { Name = "Embedded files" };
+                folder.AddChild(new Note { Text = text });
+                build.AddChild(folder);
             }
+
+            favoritesTree.TopPanel.Visibility = Visibility.Collapsed;
+            favoritesTree.ResultsList.ItemContainerStyle = treeViewItemStyle;
+            favoritesTree.ResultsList.SelectedItemChanged += ResultsList_SelectionChanged;
+            favoritesTree.ResultsList.ContextMenu = sharedTreeContextMenu;
+            favoritesTree.DisplayItems(new[] { new Note { Text = "Right-click any node and Favorite it to add it here" } });
+            favoritesTree.ResultsList.GotFocus += (s, a) => ActiveTreeView = favoritesTree.ResultsList;
 
             breadCrumb.SelectionChanged += BreadCrumb_SelectionChanged;
 
@@ -319,15 +393,19 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             filesTree.ResultsList.ItemContainerStyle = null;
             filesTree.ContextMenu = null;
             filesTree.DisplayItems(null);
+            favoritesTree.ResultsList.ItemContainerStyle = null;
             findInFilesControl.ResultsList.ItemContainerStyle = null;
             treeView.RemoveHandler(TreeViewItem.SelectedEvent, (RoutedEventHandler)TreeViewItem_Selected);
             treeView.SelectedItemChanged -= TreeView_SelectedItemChanged;
-            treeView.KeyDown -= TreeView_KeyDown;
+            treeView.KeyUp -= TreeView_KeyDown;
             treeView.GotFocus -= TreeView_GetFocus;
             treeView.ItemsSource = null;
             treeView.ItemContainerStyle = null;
             treeView.ContextMenu = null;
             centralTabControl.SelectionChanged -= CentralTabControl_SelectionChanged;
+
+            findTextBox.KeyDown -= FindTextBox_KeyDown;
+            searchLogControl.searchTextBox.KeyUp -= SearchTextBox_KeyUp;
 
             if (this.tracing.Timeline != null)
             {
@@ -351,10 +429,14 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             viewSubtreeTextItem = null;
             searchInSubtreeItem = null;
             excludeSubtreeFromSearchItem = null;
+            excludeNodeByNameFromSearch = null;
+            searchInNodeByNameItem = null;
+            searchThisNode = null;
             goToTimeLineItem = null;
             goToTracingItem = null;
             copyChildrenItem = null;
             sortChildrenItem = null;
+            filterChildrenItem = null;
             copyNameItem = null;
             copyValueItem = null;
             viewSourceItem = null;
@@ -362,10 +444,15 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             openFileItem = null;
             copyFilePathItem = null;
             preprocessItem = null;
+            searchNuGetItem = null;
             runItem = null;
             debugItem = null;
             hideItem = null;
             showTimeItem = null;
+            favoriteItem = null;
+            unfavoriteItem = null;
+            favoriteSharedItem = null;
+            unfavoriteSharedItem = null;
 
             sharedTreeContextMenu = null;
             filesTreeContextMenu = null;
@@ -376,7 +463,6 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             projectContext = null;
             SelectedTreeViewItem = null;
             sourceFileResolver = null;
-            scrollViewer = null;
             BaseNode.ClearSelectedNode();
             this.Build = null;
         }
@@ -407,7 +493,7 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             var list = filesTree.ResultsList.ItemsSource as IEnumerable<object>;
             if (list != null)
             {
-                UpdateFileVisibility(list.OfType<NamedNode>(), text.Replace(":", ""));
+                UpdateFileVisibility(list.OfType<NamedNode>(), text);
             }
         }
 
@@ -532,8 +618,6 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             "is newer than output ",
             "Property reassignment: $(",
             "out-of-date",
-            "csc $task",
-            "ResolveAssemblyReference $task",
             "$task $time",
             "$message CompilerServer failed",
             "will be compiled because",
@@ -553,7 +637,6 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
             "$additem",
             "$removeitem",
             "$metadata",
-            "$copytask",
             "$csc",
             "$rar",
             "$import",
@@ -590,22 +673,49 @@ Right-clicking a project node may show the 'Preprocess' option if the version of
 
         private void UpdateWatermark()
         {
-            string watermarkText1 = @"Type in the search box to search. Press Ctrl+F to focus the search box. Results (up to 1000) will display here.
+            string watermarkText0 = @"Type in the search box to search. Press Ctrl+F to focus the search box. Results (up to 1000) will display here.
+";
 
+            string watermarkText1 = @"
 Search for multiple words separated by space (space means AND). Enclose multiple words in double-quotes """" to search for the exact phrase. A single word in quotes means exact match (turns off substring search).
 
 Use syntax like '$property Prop' to narrow results down by item kind. Supported kinds: ";
 
             string watermarkText2 = @"Use the under(FILTER) clause to only include results where any of the nodes in the parent chain matches the FILTER. Use project(...) to filter by parent project. Examples:
- • $task csc under($project Core)
+ • $csc under($project Core)
  • Copying file project(ProjectA)
 
 Append [[$time]], [[$start]] and/or [[$end]] to show times and/or durations and sort the results by start time or duration descending (for tasks, targets and projects).
+
+Use start<""2023-11-23 14:30:54.579"", start>, end< or end> to filter events that start or end before or after a given timestamp. Timestamp needs to be in quotes.
+
+Use '$copy path' where path is a file or directory to find file copy operations involving the file or directory. `$copy substring` will search for copied files containing the substring.
+
+Use '$nuget project(MyProject.csproj) Package.Name' to search for NuGet packages (by name or version), dependencies (direct and transitive) and files coming from NuGet packages.
 
 Examples:
 ";
 
             var watermark = new TextBlock();
+            watermark.Inlines.Add(watermarkText0);
+
+            var recentSearches = SettingsService.GetRecentSearchStrings();
+            if (recentSearches.Any())
+            {
+                watermark.Inlines.Add(@"
+Recent (");
+                var clearRecentHyperlink = new Hyperlink(new Run("clear"));
+                clearRecentHyperlink.Click += (s, e) => { SettingsService.RemoveAllRecentSearchText(); UpdateWatermark(); };
+                watermark.Inlines.Add(clearRecentHyperlink);
+                watermark.Inlines.Add(@"):
+");
+
+                foreach (var recentSearch in recentSearches.Where(s => !searchExamples.Contains(s) && !nodeKinds.Contains(s)))
+                {
+                    watermark.Inlines.Add(MakeLink(recentSearch, searchLogControl));
+                }
+            }
+
             watermark.Inlines.Add(watermarkText1);
 
             bool isFirst = true;
@@ -628,19 +738,6 @@ Examples:
             foreach (var example in searchExamples)
             {
                 watermark.Inlines.Add(MakeLink(example, searchLogControl));
-            }
-
-            var recentSearches = SettingsService.GetRecentSearchStrings();
-            if (recentSearches.Any())
-            {
-                watermark.Inlines.Add(@"
-Recent:
-");
-
-                foreach (var recentSearch in recentSearches.Where(s => !searchExamples.Contains(s) && !nodeKinds.Contains(s)))
-                {
-                    watermark.Inlines.Add(MakeLink(recentSearch, searchLogControl));
-                }
             }
 
             searchLogControl.WatermarkContent = watermark;
@@ -667,7 +764,11 @@ Recent:
             if (recentSearches.Any())
             {
                 watermark.Inlines.Add(@"
-Recent:
+Recent (");
+                var clearRecentHyperlink = new Hyperlink(new Run("clear"));
+                clearRecentHyperlink.Click += (s, e) => { SettingsService.RemoveAllRecentSearchText("PropertiesAndItems"); UpdatePropertiesAndItemsWatermark(); };
+                watermark.Inlines.Add(clearRecentHyperlink);
+                watermark.Inlines.Add(@"):
 ");
 
                 foreach (var recentSearch in recentSearches)
@@ -696,6 +797,13 @@ Recent:
                     result.Add(chunk);
                 }
             }
+        }
+
+        private void SearchNuGet(IProjectOrEvaluation node)
+        {
+            string projectName = Path.GetFileName(node.ProjectFile);
+            searchLogControl.SearchText = $"$nuget project({projectName})";
+            SelectSearchTab();
         }
 
         private void Preprocess(IPreprocessable project) => preprocessedFileManager.ShowPreprocessed(project);
@@ -772,9 +880,9 @@ Recent:
         private void ContextMenu_Opened(object sender, RoutedEventArgs e)
         {
             var node = treeView.SelectedItem as BaseNode;
-            var visibility = node is NameValueNode ? Visibility.Visible : Visibility.Collapsed;
-            copyNameItem.Visibility = visibility;
-            copyValueItem.Visibility = visibility;
+            var nameValueVisibility = node is NameValueNode ? Visibility.Visible : Visibility.Collapsed;
+            copyNameItem.Visibility = nameValueVisibility;
+            copyValueItem.Visibility = nameValueVisibility;
             viewSourceItem.Visibility = CanView(node) ? Visibility.Visible : Visibility.Collapsed;
             viewFullTextItem.Visibility = HasFullText(node) ? Visibility.Visible : Visibility.Collapsed;
             openFileItem.Visibility = CanOpenFile(node) ? Visibility.Visible : Visibility.Collapsed;
@@ -782,20 +890,85 @@ Recent:
                 ? Visibility.Visible
                 : Visibility.Collapsed;
             var hasChildren = node is TreeNode t && t.HasChildren;
-            copySubtreeItem.Visibility = hasChildren ? Visibility.Visible : Visibility.Collapsed;
-            viewSubtreeTextItem.Visibility = copySubtreeItem.Visibility;
-            showTimeItem.Visibility = node is TimedNode ? Visibility.Visible : Visibility.Collapsed;
-            searchInSubtreeItem.Visibility = hasChildren && node is TimedNode ? Visibility.Visible : Visibility.Collapsed;
-            excludeSubtreeFromSearchItem.Visibility = hasChildren && node is TimedNode ? Visibility.Visible : Visibility.Collapsed;
-            goToTimeLineItem.Visibility = node is TimedNode ? Visibility.Visible : Visibility.Collapsed;
-            goToTracingItem.Visibility = node is TimedNode ? Visibility.Visible : Visibility.Collapsed;
-            copyChildrenItem.Visibility = copySubtreeItem.Visibility;
-            sortChildrenItem.Visibility = copySubtreeItem.Visibility;
+            var hasChildrenVisibility = hasChildren ? Visibility.Visible : Visibility.Collapsed;
+            copySubtreeItem.Visibility = hasChildrenVisibility;
+            viewSubtreeTextItem.Visibility = hasChildrenVisibility;
+            copyChildrenItem.Visibility = hasChildrenVisibility;
+            sortChildrenItem.Visibility = hasChildrenVisibility;
+            filterChildrenItem.Visibility = hasChildrenVisibility;
             preprocessItem.Visibility = node is IPreprocessable p && preprocessedFileManager.CanPreprocess(p) ? Visibility.Visible : Visibility.Collapsed;
+            searchNuGetItem.Visibility = node is IProjectOrEvaluation ? Visibility.Visible : Visibility.Collapsed;
             Visibility canRun = Build?.LogFilePath != null && node is Task ? Visibility.Visible : Visibility.Collapsed;
             runItem.Visibility = canRun;
             debugItem.Visibility = canRun;
             hideItem.Visibility = node is TreeNode ? Visibility.Visible : Visibility.Collapsed;
+
+            if (node is SearchableItem searchItem)
+            {
+                searchThisNode.Visibility = Visibility.Visible;
+                searchThisNode.Header = $"Search {searchItem.SearchText}";
+            }
+            else
+            {
+                searchThisNode.Visibility = Visibility.Collapsed;
+            }
+
+            bool isFavorite = IsFavorite(node);
+            favoriteItem.Visibility = !isFavorite ? Visibility.Visible : Visibility.Collapsed;
+            unfavoriteItem.Visibility = isFavorite ? Visibility.Visible : Visibility.Collapsed;
+
+            if (node is TimedNode timedNode)
+            {
+                showTimeItem.Visibility = Visibility.Visible;
+                searchInSubtreeItem.Visibility = hasChildren ? Visibility.Visible : Visibility.Collapsed;
+                excludeSubtreeFromSearchItem.Visibility = hasChildren ? Visibility.Visible : Visibility.Collapsed;
+                goToTimeLineItem.Visibility = Visibility.Visible;
+                goToTracingItem.Visibility = Visibility.Visible;
+                excludeNodeByNameFromSearch.Visibility = hasChildren ? Visibility.Visible : Visibility.Collapsed;
+                searchInNodeByNameItem.Visibility = hasChildren ? Visibility.Visible : Visibility.Collapsed;
+
+                if (excludeNodeByNameFromSearch.Visibility == Visibility.Visible)
+                {
+                    excludeNodeByNameFromSearch.Header = $"Exclude '{timedNode.Name}' from search";
+                }
+
+                if (searchInNodeByNameItem.Visibility == Visibility.Visible)
+                {
+                    searchInNodeByNameItem.Header = $"Search in '{timedNode.Name}'";
+                }
+            }
+            else
+            {
+                showTimeItem.Visibility = Visibility.Collapsed;
+                searchInSubtreeItem.Visibility = Visibility.Collapsed;
+                excludeSubtreeFromSearchItem.Visibility = Visibility.Collapsed;
+                goToTimeLineItem.Visibility = Visibility.Collapsed;
+                goToTracingItem.Visibility = Visibility.Collapsed;
+                excludeNodeByNameFromSearch.Visibility = Visibility.Collapsed;
+                searchInNodeByNameItem.Visibility = Visibility.Collapsed;
+            }
+
+            searchMenuGroup.Visibility = searchMenuGroup.Items.Cast<MenuItem>().Any(p => p.Visibility != Visibility.Collapsed) ?
+                Visibility.Visible : Visibility.Collapsed;
+
+            copyMenuGroup.Visibility = copyMenuGroup.Items.Cast<MenuItem>().Any(p => p.Visibility != Visibility.Collapsed) ?
+                Visibility.Visible : Visibility.Collapsed;
+
+            gotoMenuGroup.Visibility = gotoMenuGroup.Items.Cast<MenuItem>().Any(p => p.Visibility != Visibility.Collapsed) ?
+                Visibility.Visible : Visibility.Collapsed;
+        }
+
+        private void SharedTreeContextMenu_Opened(object sender, RoutedEventArgs e)
+        {
+            var node = ActiveTreeView.SelectedItem as BaseNode;
+            if (node == null)
+            {
+                return;
+            }
+
+            bool isFavorite = IsFavorite(node);
+            favoriteSharedItem.Visibility = !isFavorite ? Visibility.Visible : Visibility.Collapsed;
+            unfavoriteSharedItem.Visibility = isFavorite ? Visibility.Visible : Visibility.Collapsed;
         }
 
         private object FindInFiles(string searchText, int maxResults, CancellationToken cancellationToken)
@@ -961,16 +1134,14 @@ Recent:
                     Name = parts[index]
                 };
 
-                foreach (var target in GetTargets(filePath))
+                if (PlatformUtilities.HasThreads)
                 {
-                    file.AddChild(new Target
-                    {
-                        Name = target,
-                        SourceFilePath = filePath
-                    });
+                    TPLTask.Run(() => AddTargetsAsync(filePath, file));
                 }
-
-                file.SortChildren();
+                else
+                {
+                    AddTargets(filePath, file);
+                }
 
                 folder.AddChild(file);
                 return file;
@@ -991,10 +1162,60 @@ Recent:
             }
         }
 
+        private async TPLTask AddTargetsAsync(string filePath, SourceFile file)
+        {
+            var targets = GetTargets(filePath).OrderBy(t => t).ToArray();
+            if (targets.Length == 0)
+            {
+                return;
+            }
+
+            await Dispatcher.InvokeAsync(() =>
+            {
+                foreach (var target in targets)
+                {
+                    file.AddChild(new Target
+                    {
+                        Name = target,
+                        SourceFilePath = filePath
+                    });
+                }
+            });
+        }
+
+        private void AddTargets(string filePath, SourceFile file)
+        {
+            var targets = GetTargets(filePath).OrderBy(t => t).ToArray();
+            if (targets.Length == 0)
+            {
+                return;
+            }
+
+            foreach (var target in targets)
+            {
+                file.AddChild(new Target
+                {
+                    Name = target,
+                    SourceFilePath = filePath
+                });
+            }
+        }
+
+        private static HashSet<string> nonMSBuildExtensions = new(StringComparer.OrdinalIgnoreCase)
+        {
+            ".dll",
+            ".json",
+            ".rsp",
+            ".sln",
+            ".tmp",
+            ".txt",
+            ".user"
+        };
+
         private IEnumerable<string> GetTargets(string file)
         {
-            if (file.EndsWith(".sln", StringComparison.OrdinalIgnoreCase) ||
-                file.EndsWith(".dll", StringComparison.OrdinalIgnoreCase))
+            var extension = Path.GetExtension(file);
+            if (nonMSBuildExtensions.Contains(extension))
             {
                 yield break;
             }
@@ -1008,6 +1229,11 @@ Recent:
             var contentText = content.Text;
 
             if (!Utilities.LooksLikeXml(contentText))
+            {
+                yield break;
+            }
+
+            if (contentText.IndexOf("<Target", StringComparison.Ordinal) == -1)
             {
                 yield break;
             }
@@ -1051,6 +1277,7 @@ Recent:
         /// </summary>
         private bool isProcessingBreadcrumbClick = false;
         internal static TimeSpan Elapsed;
+        internal static TimeSpan PrecalculationDuration;
 
         private void BreadCrumb_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
@@ -1090,6 +1317,7 @@ Recent:
             {
                 UpdateBreadcrumb(item);
                 UpdateProjectContext(item);
+                UpdateFindContent();
             }
         }
 
@@ -1201,11 +1429,9 @@ Recent:
 
         private void BuildControl_Loaded(object sender, RoutedEventArgs e)
         {
-            scrollViewer = treeView.Template.FindName("_tv_scrollviewer_", treeView) as ScrollViewer;
-
             if (!Build.Succeeded)
             {
-                var firstError = Build.FindFirstInSubtreeIncludingSelf<Error>();
+                var firstError = Build.FirstError;
                 if (firstError != null)
                 {
                     SelectItem(firstError);
@@ -1242,20 +1468,185 @@ Recent:
 
         private void TreeView_KeyDown(object sender, KeyEventArgs args)
         {
-            if (args.Key == Key.Delete)
-            {
-                Delete();
-                args.Handled = true;
-            }
-            else if (args.Key == Key.C && args.KeyboardDevice.Modifiers == ModifierKeys.Control)
-            {
-                CopySubtree();
-                args.Handled = true;
-            }
-            else if (args.Key >= Key.A && args.Key <= Key.Z && args.KeyboardDevice.Modifiers == ModifierKeys.None)
+            if (args.Key >= Key.A && args.Key <= Key.Z && args.KeyboardDevice.Modifiers == ModifierKeys.None)
             {
                 SelectItemByKey((char)('A' + args.Key - Key.A));
                 args.Handled = true;
+            }
+        }
+
+        public bool IsFindVisible
+        {
+            get => findControl.Visibility == Visibility.Visible;
+            set
+            {
+                findControl.Visibility = value ? Visibility.Visible : Visibility.Collapsed;
+                if (value)
+                {
+                    findTextBox.Focus();
+                    UpdateFindContent();
+                }
+                else
+                {
+                    ActiveTreeView.Focus();
+                }
+            }
+        }
+
+        private TreeNode TryGetTreeNodeForFind()
+        {
+            BaseNode node = treeView.SelectedItem as BaseNode;
+            if (node is Property or Metadata)
+            {
+                node = node.Parent;
+            }
+            else if (node is Item item && !item.HasChildren)
+            {
+                node = node.Parent;
+            }
+
+            var treeNode = node as TreeNode;
+            if (treeNode != null && treeNode.HasChildren)
+            {
+                return treeNode;
+            }
+
+            return null;
+        }
+
+        private void UpdateFindContent()
+        {
+            if (!IsFindVisible)
+            {
+                return;
+            }
+
+            var treeNode = TryGetTreeNodeForFind();
+            if (treeNode != null)
+            {
+                findLabel.Content = $"Filter children of: {TextUtilities.ShortenValue(GetText(treeNode), trimPrompt: "", maxChars: 100)}";
+                if (nodeFilters.TryGetValue(treeNode, out var filter))
+                {
+                    findTextBox.Text = filter;
+                }
+                else
+                {
+                    findTextBox.Clear();
+                }
+            }
+            else
+            {
+                IsFindVisible = false;
+            }
+        }
+
+        private void SearchTextBox_KeyUp(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Key.Escape && e.KeyboardDevice.Modifiers == ModifierKeys.None)
+            {
+                if (string.IsNullOrEmpty(searchLogControl.SearchText))
+                {
+                    ActiveTreeView.Focus();
+                    e.Handled = true;
+                }
+                else
+                {
+                    searchLogControl.searchTextBox.Clear();
+                    e.Handled = true;
+                }
+            }
+        }
+
+        private void FindTextBox_KeyDown(object sender, KeyEventArgs e)
+        {
+            if (e.Handled == true)
+            {
+                return;
+            }
+
+            if (e.KeyboardDevice.Modifiers == ModifierKeys.None)
+            {
+                if (e.Key == Key.Escape)
+                {
+                    if (!string.IsNullOrEmpty(findTextBox.Text))
+                    {
+                        findTextBox.Clear();
+                    }
+                    else
+                    {
+                        IsFindVisible = false;
+                    }
+
+                    e.Handled = true;
+                }
+
+                if (e.Key == Key.Return)
+                {
+                    IsFindVisible = false;
+                    e.Handled = true;
+                }
+            }
+            else if (e.KeyboardDevice.Modifiers == ModifierKeys.Control)
+            {
+                if (e.Key == Key.F)
+                {
+                    IsFindVisible = false;
+                    FocusSearch();
+                    e.Handled = true;
+                }
+            }
+        }
+
+        private void findTextBox_TextChanged(object sender, TextChangedEventArgs e)
+        {
+            var searchText = findTextBox.Text.Trim();
+
+            var node = TryGetTreeNodeForFind();
+            if (node == null)
+            {
+                return;
+            }
+
+            ApplyFilter(node, searchText);
+        }
+
+        private readonly Dictionary<TreeNode, string> nodeFilters = new Dictionary<TreeNode, string>();
+
+        private void ApplyFilter(TreeNode node, string text)
+        {
+            if (nodeFilters.TryGetValue(node, out var existing))
+            {
+                if (existing == text)
+                {
+                    return;
+                }
+            }
+
+            var children = node.Children;
+            var view = CollectionViewSource.GetDefaultView(children);
+            if (string.IsNullOrEmpty(text))
+            {
+                view.Filter = null;
+                nodeFilters.Remove(node);
+            }
+            else
+            {
+                nodeFilters[node] = text;
+                view.Filter = o =>
+                {
+                    if (o is not BaseNode childNode)
+                    {
+                        return false;
+                    }
+
+                    var nodeText = GetText(childNode);
+                    if (nodeText != null && nodeText.IndexOf(text, StringComparison.OrdinalIgnoreCase) >= 0)
+                    {
+                        return true;
+                    }
+
+                    return false;
+                };
             }
         }
 
@@ -1286,6 +1677,11 @@ Recent:
             foreach (var item in items)
             {
                 var text = GetText(item);
+                if (text == null)
+                {
+                    continue;
+                }
+
                 if (characterMatchPrefixLength < text.Length && text.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
                 {
                     var character = text[characterMatchPrefixLength];
@@ -1305,18 +1701,11 @@ Recent:
                 items = items.Skip(1).Concat(items.Take(1));
                 goto search;
             }
+        }
 
-            string GetText(BaseNode node)
-            {
-                if (node is IHasTitle hasTitle)
-                {
-                    return hasTitle.Title ?? "";
-                }
-                else
-                {
-                    return node.ToString() ?? "";
-                }
-            }
+        private string GetText(BaseNode node)
+        {
+            return node.Title ?? node.ToString();
         }
 
         public void FocusSearch()
@@ -1352,7 +1741,8 @@ Recent:
 
         public void Copy()
         {
-            var treeNode = treeView.SelectedItem;
+            var tree = ActiveTreeView;
+            var treeNode = tree?.SelectedItem;
             if (treeNode != null)
             {
                 var text = treeNode.ToString();
@@ -1362,7 +1752,7 @@ Recent:
 
         public void CopySubtree(TreeView tree = null)
         {
-            tree = tree ?? ActiveTreeView;
+            tree ??= ActiveTreeView;
             if (tree == null)
             {
                 return;
@@ -1390,6 +1780,115 @@ Recent:
             {
                 var text = timedNode.GetTimeAndDurationText(fullPrecision: true);
                 DisplayText(text, timedNode.ToString());
+            }
+        }
+
+        private readonly HashSet<BaseNode> favorites = new HashSet<BaseNode>();
+
+        public void AddToFavorites()
+        {
+            var node = ActiveTreeView?.SelectedItem as BaseNode;
+            if (node != null)
+            {
+                if (node is ProxyNode proxy)
+                {
+                    node = proxy.Original ?? node;
+                }
+
+                if (favorites.Add(node))
+                {
+                    RefreshFavorites();
+                }
+            }
+        }
+
+        public void RemoveFromFavorites()
+        {
+            var node = ActiveTreeView?.SelectedItem as BaseNode;
+            if (node != null)
+            {
+                if (node is ProxyNode proxy)
+                {
+                    node = proxy.Original ?? node;
+                }
+
+                if (favorites.Remove(node))
+                {
+                    RefreshFavorites();
+                }
+            }
+        }
+
+        public bool IsFavorite(BaseNode node)
+        {
+            if (node is ProxyNode proxy)
+            {
+                node = proxy.Original ?? node;
+            }
+
+            return favorites.Contains(node);
+        }
+
+        public void RefreshFavorites()
+        {
+            var list = favorites.OrderBy(f =>
+            {
+                if (f is TimedNode timed)
+                {
+                    return timed.Index;
+                }
+
+                return 0;
+            }).Select(f =>
+            {
+                var searchResult = new SearchResult(f);
+                return searchResult;
+            }).ToArray();
+
+            var tree = ResultTree.BuildResultTree(
+                list,
+                addDuration: false,
+                addWhenNoResults: () => new Note { Text = "Right-click any node and Favorite it to add it here" });
+
+            SortByIndex(tree);
+
+            favoritesTree.DisplayItems(tree.Children);
+        }
+
+        private static int CompareByIndex(BaseNode l, BaseNode r)
+        {
+            if (l == r)
+            {
+                return 0;
+            }
+
+            if (l is null || r is null)
+            {
+                return -1;
+            }
+
+            if (l is TimedNode timedLeft && r is TimedNode timedRight)
+            {
+                return timedLeft.Index - timedRight.Index;
+            }
+
+            return 0;
+        }
+
+        private void SortByIndex(TreeNode node)
+        {
+            node.SortChildren(CompareByIndex);
+            SortByIndex(node.Children);
+        }
+
+        private void SortByIndex(IList<BaseNode> list)
+        {
+            foreach (var child in list)
+            {
+                if (child is TreeNode childNode)
+                {
+                    SortByIndex(childNode);
+                }
             }
         }
 
@@ -1428,11 +1927,38 @@ Recent:
             }
         }
 
+        public void SearchInNodeByName()
+        {
+            if (treeView.SelectedItem is TimedNode treeNode)
+            {
+                searchLogControl.SearchText += $" under(${treeNode.TypeName} {treeNode.Name})";
+                SelectSearchTab();
+            }
+        }
+
+        public void SearchThisNode()
+        {
+            if (treeView.SelectedItem is SearchableItem searchNode)
+            {
+                searchLogControl.SearchText = searchNode.SearchText;
+                SelectSearchTab();
+            }
+        }
+
         public void ExcludeSubtreeFromSearch()
         {
             if (treeView.SelectedItem is TimedNode treeNode)
             {
                 searchLogControl.SearchText += $" notunder(${treeNode.Index})";
+                SelectSearchTab();
+            }
+        }
+
+        public void ExcludeNodeByNameFromSearch()
+        {
+            if (treeView.SelectedItem is TimedNode treeNode)
+            {
+                searchLogControl.SearchText += $" notunder(${treeNode.TypeName} {treeNode.Name})";
                 SelectSearchTab();
             }
         }
@@ -1483,9 +2009,14 @@ Recent:
             }
         }
 
+        public void FilterChildren()
+        {
+            IsFindVisible = !IsFindVisible;
+        }
+
         private void CopyAll(TreeView tree = null)
         {
-            tree = tree ?? ActiveTreeView;
+            tree ??= ActiveTreeView;
             if (tree == null)
             {
                 return;
@@ -1499,6 +2030,11 @@ Recent:
                 if (!text.Contains("\n"))
                 {
                     sb.AppendLine();
+                }
+
+                if (sb.Length > Microsoft.Build.Logging.StructuredLogger.StringWriter.MaxStringLength)
+                {
+                    break;
                 }
             }
 
@@ -1518,6 +2054,11 @@ Recent:
             {
                 item.VisitAllChildren<BaseNode>(s =>
                 {
+                    if (sb.Length > Microsoft.Build.Logging.StructuredLogger.StringWriter.MaxStringLength)
+                    {
+                        return;
+                    }
+
                     if (s is SourceFile file && !string.IsNullOrEmpty(file.SourceFilePath))
                     {
                         sb.AppendLine(file.SourceFilePath);
@@ -1590,20 +2131,63 @@ Recent:
 
         private void OnItemKeyDown(object sender, KeyEventArgs args)
         {
-            if (args.Key == Key.Space || args.Key == Key.Return)
+            if (args.Handled)
             {
-                var treeNode = GetNode(args);
-                if (treeNode != null)
+                return;
+            }
+
+            if (args.KeyboardDevice.Modifiers == ModifierKeys.None)
+            {
+                if (args.Key == Key.Delete)
                 {
-                    args.Handled = Invoke(treeNode) || ViewFullText(treeNode);
+                    Delete();
+                    args.Handled = true;
+                }
+
+                if (args.Key == Key.Space || args.Key == Key.Return)
+                {
+                    var treeNode = GetNode(args);
+                    if (treeNode != null)
+                    {
+                        args.Handled = Invoke(treeNode) || ViewFullText(treeNode);
+                    }
+                }
+
+                if (args.Key == Key.Escape)
+                {
+                    if (IsFindVisible)
+                    {
+                        IsFindVisible = false;
+                        args.Handled = true;
+                    }
+                    else if (documentWell.IsVisible)
+                    {
+                        documentWell.Hide();
+                        args.Handled = true;
+                    }
                 }
             }
 
-            if (args.Key == Key.Escape)
+            if (args.KeyboardDevice.Modifiers == ModifierKeys.Control)
             {
-                if (documentWell.IsVisible)
+                if (args.Key == Key.C)
                 {
-                    documentWell.Hide();
+                    Copy();
+                    args.Handled = true;
+                }
+
+                if (args.Key == Key.F)
+                {
+                    if (IsFindVisible)
+                    {
+                        IsFindVisible = false;
+                        args.Handled = true;
+                    }
+                    else if (TryGetTreeNodeForFind() != null)
+                    {
+                        IsFindVisible = true;
+                        args.Handled = true;
+                    }
                 }
             }
         }
@@ -1624,14 +2208,6 @@ Recent:
             }
         }
 
-        private void OnPreviewMouseRightButtonDown(object sender, MouseButtonEventArgs args)
-        {
-            if (sender is TreeViewItem treeViewItem)
-            {
-                treeViewItem.IsSelected = true;
-            }
-        }
-
         private bool CanView(BaseNode node)
         {
             return node is AbstractDiagnostic
@@ -1644,6 +2220,7 @@ Recent:
         private bool HasFullText(BaseNode node)
         {
             return (node is NameValueNode nvn && nvn.IsValueShortened)
+                || (node is NamedNode nn && nn.IsNameShortened)
                 || (node is TextNode tn && tn.IsTextShortened);
         }
 
@@ -1664,7 +2241,9 @@ Recent:
                 case NameValueNode nameValueNode when nameValueNode.IsValueShortened:
                     return DisplayText(nameValueNode.Value, nameValueNode.Name);
                 case TextNode textNode when textNode.IsTextShortened:
-                    return DisplayText(textNode.Text, textNode.Name ?? textNode.GetType().Name);
+                    return DisplayText(textNode.Text, textNode.ShortenedText ?? textNode.TypeName);
+                case NamedNode namedNode when namedNode.IsNameShortened:
+                    return DisplayText(namedNode.Name, namedNode.ShortenedName ?? namedNode.TypeName);
                 default:
                     return false;
             }
@@ -1702,6 +2281,14 @@ Recent:
                         return DisplayAddRemoveItem(addItem.Parent, addItem.LineNumber ?? 0);
                     case RemoveItem removeItem:
                         return DisplayAddRemoveItem(removeItem.Parent, removeItem.LineNumber ?? 0);
+                    case Item item when item.Parent is AddItem parentAddItem && parentAddItem.Name == "EmbedInBinlog":
+                        return DisplayEmbeddedFile(item);
+                    case Item item when
+                        item.Parent == null &&
+                        searchLogControl.SearchText.Contains("$copy") &&
+                        searchLogControl.ResultsList.ItemsSource is IEnumerable<BaseNode> results &&
+                        results.Contains(item):
+                        return SearchForFullPath(item.Text);
                     case IHasSourceFile hasSourceFile when hasSourceFile.SourceFilePath != null:
                         int line = 0;
                         var hasLine = hasSourceFile as IHasLineNumber;
@@ -1715,6 +2302,11 @@ Recent:
                         {
                             // TODO: https://github.com/KirillOsenkov/MSBuildStructuredLog/issues/392
                             evaluation = node.GetNearestParentOrSelf<ProjectEvaluation>();
+
+                            if (evaluation == null && node is Project project)
+                            {
+                                evaluation = Build.FindEvaluation(project.EvaluationId);
+                            }
                         }
 
                         return DisplayFile(hasSourceFile.SourceFilePath, line, evaluation: evaluation);
@@ -1732,6 +2324,35 @@ Recent:
             return false;
         }
 
+        private bool SearchForFullPath(string filePath)
+        {
+            var text = searchLogControl.SearchText;
+            var matcher = new NodeQueryMatcher(text);
+            if (matcher.Terms.Count == 1 &&
+                matcher.Terms[0].Word is string substring &&
+                filePath.IndexOf(substring, StringComparison.OrdinalIgnoreCase) != -1)
+            {
+                text = text.Replace(substring, filePath);
+                searchLogControl.SearchText = text;
+                return true;
+            }
+
+            return false;
+        }
+
+        private bool DisplayEmbeddedFile(Item item)
+        {
+            string path = item.Text;
+
+            var candidates = sourceFileResolver.ArchiveFile.FindFileNames(path).ToArray();
+            if (candidates.Length == 1)
+            {
+                return DisplayFile(candidates[0]);
+            }
+
+            return false;
+        }
+
         public bool DisplayFile(string sourceFilePath, int lineNumber = 0, int column = 0, ProjectEvaluation evaluation = null)
         {
             var text = sourceFileResolver.GetSourceFileText(sourceFilePath);
@@ -1740,7 +2361,7 @@ Recent:
                 return false;
             }
 
-            string preprocessableFilePath = Utilities.InsertMissingDriveSeparator(sourceFilePath);
+            string preprocessableFilePath = sourceFilePath;
 
             Action preprocess = null;
             if (evaluation != null)
@@ -1848,7 +2469,16 @@ Recent:
 
         public IEnumerable BuildResultTree(object resultsObject, bool moreAvailable = false)
         {
-            var folder = ResultTree.BuildResultTree(resultsObject, moreAvailable, Elapsed);
+            return BuildResultTree(resultsObject, moreAvailable, addDuration: true);
+        }
+
+        public IEnumerable BuildResultTree(object resultsObject, bool moreAvailable = false, bool addDuration = true)
+        {
+            var folder = ResultTree.BuildResultTree(
+                resultsObject,
+                Elapsed,
+                addDuration: addDuration,
+                addWhenNoResults: () => new Message { Text = "No results found." });
 
             if (moreAvailable)
             {
@@ -1873,36 +2503,6 @@ Recent:
             return folder.Children;
         }
 
-        private void TreeViewItem_RequestBringIntoView(object sender, RequestBringIntoViewEventArgs e)
-        {
-            if (scrollViewer == null)
-            {
-                return;
-            }
-
-            var treeViewItem = (TreeViewItem)sender;
-            var treeView = (TreeView)typeof(TreeViewItem).GetProperty("ParentTreeView", BindingFlags.Instance | BindingFlags.NonPublic).GetValue(treeViewItem);
-
-            if (PresentationSource.FromDependencyObject(treeViewItem) == null)
-            {
-                // the item might have disconnected by the time we run this
-                return;
-            }
-
-            Point topLeftInTreeViewCoordinates = treeViewItem.TransformToAncestor(treeView).Transform(new Point(0, 0));
-            var treeViewItemTop = topLeftInTreeViewCoordinates.Y;
-            if (treeViewItemTop < 0
-                || treeViewItemTop + treeViewItem.ActualHeight > scrollViewer.ViewportHeight
-                || treeViewItem.ActualHeight > scrollViewer.ViewportHeight)
-            {
-                // if the item is not visible or too "tall", don't do anything; let them scroll it into view
-                return;
-            }
-
-            // if the item is already fully within the viewport vertically, disallow horizontal scrolling
-            e.Handled = true;
-        }
-
         private void TreeViewItem_Selected(object sender, RoutedEventArgs e)
         {
             SelectedTreeViewItem = e.OriginalSource as TreeViewItem;
@@ -1923,8 +2523,6 @@ Recent:
 
             var recordStats = BinlogStats.Calculate(this.LogFilePath);
             var records = recordStats.CategorizedRecords;
-
-            Build.Unseal();
 
             statsRoot = DisplayRecordStats(records, Build);
 
@@ -2087,9 +2685,13 @@ Recent:
         private Folder DisplayRecordStats(BinlogStats.RecordsByType stats, TreeNode parent, string titlePrefix = "")
         {
             var node = parent.GetOrCreateNodeWithName<Folder>(titlePrefix + stats.ToString());
-            foreach (var records in stats.CategorizedRecords)
+
+            if (stats.CategorizedRecords != null)
             {
-                DisplayRecordStats(records, node);
+                foreach (var records in stats.CategorizedRecords)
+                {
+                    DisplayRecordStats(records, node);
+                }
             }
 
             var top = stats.Records.Take(300).ToArray();
