@@ -696,7 +696,13 @@ namespace StructuredLogViewer.Controls
 
         private Panel CreatePanelForNodeDivider(bool showTime)
         {
-            var timeWidth = ConvertTimeToPixel(GlobalEndTime - GlobalStartTime);
+            var endTime = GlobalEndTime;
+            if (endTime < GlobalStartTime)
+            {
+                endTime = GlobalStartTime;
+            }
+
+            var timeWidth = ConvertTimeToPixel(endTime - GlobalStartTime);
 
             bool fiveSeconds = false;
             double gapWidth;
@@ -734,6 +740,10 @@ namespace StructuredLogViewer.Controls
                     {
                         var textBlock = new TextBlock();
                         textBlock.Text = $"{i}s";
+
+                        // Set these to make TextBlock scale like TrueType.
+                        textBlock.SnapsToDevicePixels = true;
+                        TextOptions.SetTextFormattingMode(textBlock, TextFormattingMode.Ideal);
 
                         // add textHeight/2 pixels of front padding
                         Canvas.SetLeft(textBlock, textHeight / 2 + i * OneSecondPixelWidth);
