@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.ComponentModel;
 using System.Windows;
 using System.Windows.Media;
@@ -87,6 +87,7 @@ namespace StructuredLogViewer
                 SetResource("ImportStroke", Brushes.Sienna);
                 SetResource("NoImportStroke", GetBrush("#FF0000"));
                 SetResource("NoImportFill", Brushes.BlanchedAlmond);
+                SetResource("NuGet", Brushes.DeepSkyBlue);
             }
             else if (UseDarkTheme)
             {
@@ -115,6 +116,7 @@ namespace StructuredLogViewer
                 SetResource("NoImportFill", GetBrush("#474138"));
                 SetResource("TargetStroke", GetBrush("#C0A0F0"));
                 SetResource("AddItemStroke", GetBrush("#40B0B0"));
+                SetResource("NuGet", Brushes.DeepSkyBlue);
                 SetResource("\u01D6", GetBrush("#C0C0C0"));
                 return;
             }
@@ -129,6 +131,7 @@ namespace StructuredLogViewer
                 SetResource("NoImportFill", Brushes.BlanchedAlmond);
                 SetResource("TargetStroke", Brushes.MediumPurple);
                 SetResource("AddItemStroke", Brushes.Teal);
+                SetResource("NuGet", GetBrush("#004880"));
                 SetResource("\u01D6", GetBrush("#595959"));
             }
 
