@@ -5,10 +5,16 @@ namespace StructuredLogger.Tests
 {
     public class ItemGroupParserTests
     {
+        public ItemGroupParserTests()
+        {
+            Strings.Initialize("en-US");
+            Assert.Equal("en-US", Strings.ResourceSet.Culture);
+            Assert.NotNull(Strings.OutputItemsMessagePrefix);
+        }
+
         [Fact]
         public void AddItemWithMultilineMetadata()
         {
-            Strings.Initialize();
             var result = ItemGroupParser.ParsePropertyOrItemList(@"Added Item(s): 
     Link=
         tmp
@@ -23,7 +29,6 @@ namespace StructuredLogger.Tests
         [Fact]
         public void ParseSuggestedBindingRedirectsMetadata()
         {
-            Strings.Initialize();
             var parameter = ItemGroupParser.ParsePropertyOrItemList(@"Output Item(s): 
     SuggestedBindingRedirects=
         Microsoft.Build, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a
@@ -49,8 +54,6 @@ namespace StructuredLogger.Tests
         [Fact]
         public void ParseThereWasAConflict()
         {
-            Strings.Initialize();
-
             var lines = @"""System.IO.Compression, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089"" was chosen because it was primary and ""System.IO.Compression, Version=4.1.1.0, Culture=neutral, PublicKeyToken=b77a5c561934e089"" was not.
 References which depend on ""System.IO.Compression, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089"" [C:\Program Files (x86)\System.IO.Compression.dll].
     C:\Program Files (x86)\System.IO.Compression.dll
@@ -120,7 +123,6 @@ References which depend on ""System.IO.Compression, Version=4.1.1.0, Culture=neu
         [Fact]
         public void ParseMultilineMetadata()
         {
-            Strings.Initialize();
             var parameter = ItemGroupParser.ParsePropertyOrItemList(@"Added Item(s): 
     _ProjectsFiles=
         Project1
