@@ -1,7 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using static Microsoft.Build.Logging.StructuredLogger.BinLogReader;
 
 namespace Microsoft.Build.Logging.StructuredLogger
 {
@@ -26,8 +28,11 @@ namespace Microsoft.Build.Logging.StructuredLogger
         }
 
         public static Build ReadBuild(string filePath) => ReadBuild(filePath, progress: null);
+        public static Build ReadBuild(string filePath, Progress progress) => ReadBuild(filePath, progress, readerSettings: null);
+        public static Build ReadBuild(string filePath, ReaderSettings readerSettings)
+            => ReadBuild(filePath, progress: null, readerSettings);
 
-        public static Build ReadBuild(string filePath, Progress progress)
+        public static Build ReadBuild(string filePath, Progress progress, ReaderSettings readerSettings)
         {
             using (var stream = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.Read))
             {
@@ -38,7 +43,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     projectImportsArchive = File.ReadAllBytes(projectImportsZipFile);
                 }
 
-                var build = ReadBuild(stream, progress, projectImportsArchive);
+                var build = ReadBuild(stream, progress, projectImportsArchive, readerSettings);
                 build.LogFilePath = filePath;
                 return build;
             }
@@ -47,14 +52,20 @@ namespace Microsoft.Build.Logging.StructuredLogger
         public static Build ReadBuild(Stream stream, byte[] projectImportsArchive = null)
             => ReadBuild(stream, progress: null, projectImportsArchive: projectImportsArchive);
 
-        internal const bool ReuseBinlogStrings = false;
-
         public static Build ReadBuild(Stream stream, Progress progress, byte[] projectImportsArchive = null)
-        {
-            var eventSource = new BinLogReader();
+            => ReadBuild(stream, progress, projectImportsArchive, readerSettings: null);
 
+        public static Build ReadBuild(
+            Stream stream,
+            Progress progress,
+            byte[] projectImportsArchive = null,
+            ReaderSettings readerSettings = null)
+        {
             Build build = null;
             IEnumerable<string> strings = null;
+            readerSettings ??= ReaderSettings.Default;
+
+            var eventSource = new BinLogReader();
 
             eventSource.OnBlobRead += (kind, bytes) =>
             {
@@ -74,6 +85,23 @@ namespace Microsoft.Build.Logging.StructuredLogger
             {
                 strings = s;
             };
+            int[] errorByType = new int[Enum.GetValues(typeof(ReaderErrorType)).Length];
+            eventSource.RecoverableReadError += eArg =>
+            {
+                if (readerSettings.UnknownDataBehavior == UnknownDataBehavior.ThrowException)
+                {
+                    throw new Exception($"Unknown data encountered in the log file ({eArg.ErrorType}-{eArg.RecordKind}): {eArg.GetFormattedMessage()}");
+                }
+
+                if (readerSettings.UnknownDataBehavior == UnknownDataBehavior.Ignore)
+                {
+                    return;
+                }
+
+                errorByType[(int)eArg.ErrorType] += 1;
+            };
+
+            //build.AddChild(new );
 
             StructuredLogger.SaveLogToDisk = false;
             StructuredLogger.CurrentBuild = null;
@@ -82,6 +110,11 @@ namespace Microsoft.Build.Logging.StructuredLogger
             structuredLogger.Initialize(eventSource);
 
             build = structuredLogger.Construction.Build;
+
+            if (stream is FileStream)
+            {
+                structuredLogger.Construction.IsLargeBinlog = stream.Length > 100_000_000;
+            }
 
             eventSource.OnFileFormatVersionRead += fileFormatVersion =>
             {
@@ -92,10 +125,6 @@ namespace Microsoft.Build.Logging.StructuredLogger
                 {
                     build.StringTable.NormalizeLineEndings = false;
                     build.StringTable.HasDeduplicatedStrings = true;
-                    if (ReuseBinlogStrings)
-                    {
-                        build.StringTable.DisableDeduplication = true;
-                    }
                 }
             };
 
@@ -107,17 +136,9 @@ namespace Microsoft.Build.Logging.StructuredLogger
 
             if (strings != null)
             {
-                if (ReuseBinlogStrings)
-                {
-                    // since strings are already deduplicated in the file, no need to do it again
-                    build.StringTable.SetStrings(strings);
-                }
-                else
-                {
-                    // intern all strings in one fell swoop here instead of interning multiple times
-                    // one by one when processing task parameters
-                    build.StringTable.Intern(strings);
-                }
+                // intern all strings in one fell swoop here instead of interning multiple times
+                // one by one when processing task parameters
+                build.StringTable.Intern(strings);
             }
 
             structuredLogger.Shutdown();
@@ -139,6 +160,24 @@ namespace Microsoft.Build.Logging.StructuredLogger
             // strings = build.StringTable.Instances.OrderBy(s => s).ToArray();
 
             // Serialization.WriteStringsToFile(@"C:\temp\1.txt", strings.ToArray());
+
+            build.WaitForBackgroundTasks();
+
+            if (errorByType.Any(i => i != 0))
+            {
+                string summary = string.Join(", ", errorByType.Where((count, index) => count > 0).Select((count, index) => $"{((ReaderErrorType)index)}: {count}"));
+                string message = $"Skipped some data unknown to this version of Viewer. {errorByType.Sum()} case{(errorByType.Sum() > 1 ? "s" : string.Empty)} encountered ({summary}).";
+
+                TreeNode node = readerSettings.UnknownDataBehavior switch
+                {
+                    UnknownDataBehavior.Error => new Error() { Text = message },
+                    UnknownDataBehavior.Warning => new Warning() { Text = message },
+                    UnknownDataBehavior.Message => new CriticalBuildMessage() { Text = message },
+                    _ => throw new ArgumentOutOfRangeException(nameof(readerSettings.UnknownDataBehavior), readerSettings.UnknownDataBehavior, "Unexpected value")
+                };
+
+                build.AddChild(node);
+            }
 
             return build;
         }
