@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -20,14 +20,13 @@ namespace Microsoft.Build.Logging.StructuredLogger
 
         private struct NameValueRecord
         {
-            public (int keyIndex, int valueIndex)[] Array;
             public IDictionary<string, string> Dictionary;
         }
 
-        private IDictionary<string, string> CreateDictionary((int keyIndex, int valueIndex)[] list)
+        private IDictionary<string, string> CreateDictionary(List<(int keyIndex, int valueIndex)> list)
         {
-            var dictionary = new ArrayDictionary<string, string>(list.Length);
-            for (int i = 0; i < list.Length; i++)
+            var dictionary = new ArrayDictionary<string, string>(list.Count);
+            for (int i = 0; i < list.Count; i++)
             {
                 string key = GetStringFromRecord(list[i].keyIndex);
                 string value = GetStringFromRecord(list[i].valueIndex);
@@ -62,9 +61,64 @@ namespace Microsoft.Build.Logging.StructuredLogger
             return FormatResourceStringIgnoreCodeAndKeyword(succeeded ? "Done building project \"{0}\"." : "Done building project \"{0}\" -- FAILED.", Path.GetFileName(projectFile));
         }
 
+        private readonly Dictionary<(string, string, string, string), string> propertyReassignmentCache =
+            new Dictionary<(string, string, string, string), string>();
+
         private string GetPropertyReassignmentMessage(string propertyName, string newValue, string previousValue, string location)
         {
-            return FormatResourceStringIgnoreCodeAndKeyword("Property reassignment: $({0})=\"{1}\" (previous value: \"{2}\") at {3}", propertyName, newValue, previousValue, location);
+            var key = (propertyName, newValue, previousValue, location);
+            if (!propertyReassignmentCache.TryGetValue(key, out var result))
+            {
+                result = FormatResourceStringIgnoreCodeAndKeyword(Strings.PropertyReassignment, propertyName, newValue, previousValue, location);
+                propertyReassignmentCache[key] = result;
+            }
+
+            return result;
+        }
+
+        private BuildEventArgs SynthesizePropertyReassignment(BuildEventArgsFields fields)
+        {
+            string propertyName = fields.Arguments[0] as string;
+            string newValue = fields.Arguments[1] as string;
+            string previousValue = fields.Arguments[2] as string;
+            string location = fields.Arguments[3] as string;
+            string message = GetPropertyReassignmentMessage(propertyName, newValue, previousValue, location);
+
+            var e = new PropertyReassignmentEventArgs(
+                propertyName,
+                previousValue,
+                newValue,
+                location,
+                message,
+                fields.HelpKeyword,
+                fields.SenderName,
+                fields.Importance);
+            SetCommonFields(e, fields);
+
+            return e;
+        }
+
+        private bool sawCulture;
+
+        private void OnMessageRead(BuildMessageEventArgs args)
+        {
+            if (sawCulture)
+            {
+                return;
+            }
+
+            if (args.SenderName == "BinaryLogger" &&
+                args.Message is string message &&
+                message.StartsWith("CurrentUICulture", StringComparison.Ordinal))
+            {
+                sawCulture = true;
+                var kvp = TextUtilities.ParseNameValue(message);
+                string culture = kvp.Value;
+                if (!string.IsNullOrEmpty(culture))
+                {
+                    Strings.Initialize(culture);
+                }
+            }
         }
 
         private string GetTargetStartedMessage(string projectFile, string targetFile, string parentTarget, string targetName)
