@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Text;
+using System.Text.RegularExpressions;
 using Microsoft.Build.Framework;
 using Microsoft.Build.Internal;
 
@@ -11,6 +12,8 @@ namespace Microsoft.Build.Logging.StructuredLogger
     {
         private readonly Construction construction;
         private readonly StringCache stringTable;
+        private readonly ImportTreeAnalyzer importTreeAnalyzer;
+        private int fileFormatVersion = 0;
 
         public StringBuilder DetailedSummary { get; } = new StringBuilder();
 
@@ -18,6 +21,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
         {
             this.construction = construction;
             this.stringTable = stringTable;
+            this.importTreeAnalyzer = new ImportTreeAnalyzer(stringTable);
         }
 
         private string Intern(string text) => stringTable.Intern(text);
@@ -27,6 +31,11 @@ namespace Microsoft.Build.Logging.StructuredLogger
             if (args == null)
             {
                 return;
+            }
+
+            if (fileFormatVersion == 0)
+            {
+                fileFormatVersion = construction.Build.FileFormatVersion;
             }
 
             if (args is TaskParameterEventArgs taskParameter)
@@ -39,8 +48,17 @@ namespace Microsoft.Build.Logging.StructuredLogger
                 ProcessProjectImported(projectImported);
                 return;
             }
+            else if (args is TaskCommandLineEventArgs taskArgs)
+            {
+                if (AddCommandLine(taskArgs))
+                {
+                    return;
+                }
+            }
 
-            var message = args.Message;
+            // This realizes the expensive message string from LazyFormattedBuildEventArgs
+            string message = args.Message;
+
             if (string.IsNullOrEmpty(message))
             {
                 return;
@@ -49,7 +67,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
             var buildEventContext = args.BuildEventContext;
             if (buildEventContext != null && buildEventContext.TaskId != BuildEventContext.InvalidTaskId)
             {
-                if (message.StartsWith(Strings.OutputItemsMessagePrefix, StringComparison.Ordinal))
+                if (fileFormatVersion < 11 && message.StartsWith(Strings.OutputItemsMessagePrefix, StringComparison.Ordinal))
                 {
                     var task = GetTask(args);
 
@@ -85,27 +103,23 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     folder.AddChild(parameter);
                     return;
                 }
-
-                if (args is TaskCommandLineEventArgs taskArgs)
-                {
-                    if (AddCommandLine(taskArgs))
-                    {
-                        return;
-                    }
-                }
             }
             else if (buildEventContext != null && buildEventContext.TargetId != BuildEventContext.InvalidTargetId)
             {
-                if (message.StartsWith(Strings.ItemGroupIncludeMessagePrefix, StringComparison.Ordinal))
+                // since version 11 these have been replaced with TaskParameterEventArgs
+                if (fileFormatVersion < 11)
                 {
-                    AddItemGroup(args, message, Strings.ItemGroupIncludeMessagePrefix, new AddItem());
-                    return;
-                }
+                    if (message.StartsWith(Strings.ItemGroupIncludeMessagePrefix, StringComparison.Ordinal))
+                    {
+                        AddItemGroup(args, message, Strings.ItemGroupIncludeMessagePrefix, new AddItem());
+                        return;
+                    }
 
-                if (message.StartsWith(Strings.ItemGroupRemoveMessagePrefix, StringComparison.Ordinal))
-                {
-                    AddItemGroup(args, message, Strings.ItemGroupRemoveMessagePrefix, new RemoveItem());
-                    return;
+                    if (message.StartsWith(Strings.ItemGroupRemoveMessagePrefix, StringComparison.Ordinal))
+                    {
+                        AddItemGroup(args, message, Strings.ItemGroupRemoveMessagePrefix, new RemoveItem());
+                        return;
+                    }
                 }
 
                 if (message.StartsWith(Strings.PropertyGroupMessagePrefix, StringComparison.Ordinal))
@@ -114,8 +128,8 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     return;
                 }
 
-                // A task from assembly message (parses out the task name and assembly path).
-                var match = Strings.UsingTask(message);
+                string rawMessage = Reflector.GetMessage(args);
+                var match = Strings.UsingTask(message, rawMessage);
                 if (match.Success)
                 {
                     construction.SetTaskAssembly(
@@ -143,7 +157,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
 
         private void ProcessProjectImported(ProjectImportedEventArgs args)
         {
-            var import = ImportTreeAnalyzer.TryGetImportOrNoImport(args, stringTable);
+            var import = importTreeAnalyzer.TryGetImportOrNoImport(args);
             if (import == null)
             {
                 return;
@@ -177,6 +191,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
 
                 string folderName = isOutput ? Strings.OutputItems : Strings.Parameters;
                 parent = task.GetOrCreateNodeWithName<Folder>(folderName);
+                parent.DisableChildrenCache = true;
 
                 node = CreateParameterNode(itemType, items, isOutput);
             }
@@ -214,6 +229,8 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     {
                         itemType = Strings.Outputs;
                     }
+
+                    named.DisableChildrenCache = true;
                 }
 
                 named.Name = itemType;
@@ -248,6 +265,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
             else
             {
                 parent = new Parameter { Name = itemName };
+                parent.DisableChildrenCache = true;
             }
 
             AddItems(items, parent);
@@ -256,6 +274,11 @@ namespace Microsoft.Build.Logging.StructuredLogger
         }
 
         private void AddItems(IEnumerable items, TreeNode parent)
+        {
+            construction.Build.RunInBackground(() => AddItemsCore(items, parent));
+        }
+
+        private void AddItemsCore(IEnumerable items, TreeNode parent)
         {
             if (items is ICollection collection)
             {
@@ -415,7 +438,8 @@ namespace Microsoft.Build.Logging.StructuredLogger
                             return;
                         }
                     }
-                    else if (string.Equals(task.Name, "RestoreTask", StringComparison.OrdinalIgnoreCase))
+                    else if (string.Equals(task.Name, "RestoreTask", StringComparison.OrdinalIgnoreCase) ||
+                        string.Equals(task.Name, "RestoreTaskEx", StringComparison.OrdinalIgnoreCase))
                     {
                         if (ProcessRestoreTask(task, ref parent, message))
                         {
@@ -464,7 +488,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
                         args2.BuildEventContext = args.BuildEventContext;
                         args2.SkipReason = targetSkipReason;
                         args2.OriginallySucceeded = targetSkipReason != TargetSkipReason.PreviouslyBuiltUnsuccessfully;
-                        Reflector.BuildEventArgs_timestamp.SetValue(args2, args.Timestamp);
+                        Reflector.SetTimestamp(args2, args.Timestamp);
                         construction.TargetSkipped(args2);
                         return;
                     }
@@ -481,7 +505,10 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     parent = evaluation;
                 }
 
-                if (args is PropertyReassignmentEventArgs || (message.Contains(Strings.PropertyReassignment) && Strings.PropertyReassignmentRegex.IsMatch(message)))
+                Match match = null;
+                PropertyReassignmentEventArgs propertyReassignment = args as PropertyReassignmentEventArgs;
+                if (propertyReassignment != null ||
+                    ((match = Strings.PropertyReassignmentRegex.Match(message)) != null && match.Success))
                 {
                     TimedNode properties;
                     if (evaluation != null)
@@ -493,16 +520,21 @@ namespace Microsoft.Build.Logging.StructuredLogger
                         properties = parent.GetOrCreateNodeWithName<TimedNode>(Strings.PropertyReassignmentFolder, addAtBeginning: true);
                     }
 
-                    var propertyName = Strings.GetPropertyName(message);
+                    var propertyName = propertyReassignment != null ?
+                        propertyReassignment.PropertyName :
+                        match != null ?
+                            match.Groups["Name"].Value :
+                            Strings.GetPropertyName(message);
+
                     parent = properties.GetOrCreateNodeWithName<Folder>(propertyName);
                 }
-                else if (parent == evaluation && parent.FindChild<Message>(message) != null)
+                else if (parent == evaluation && !evaluation.MessageTexts.Add(message))
                 {
                     // avoid duplicate messages
                     return;
                 }
             }
-            else if (args.Message.StartsWith(Strings.NodesReusal, StringComparison.Ordinal))
+            else if (message.StartsWith(Strings.NodesReusal, StringComparison.Ordinal))
             {
                 parent = construction.Build.GetOrCreateNodeWithName<Folder>(Strings.NodesManagementNode);
             }
@@ -520,7 +552,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
 
                     parent = construction.EvaluationFolder;
                 }
-                else if (construction.Build.FileFormatVersion < 9 && message.Contains(Strings.PropertyReassignment) && Strings.PropertyReassignmentRegex.IsMatch(message))
+                else if (construction.Build.FileFormatVersion < 9 && Strings.PropertyReassignmentRegex.IsMatch(message))
                 {
                     if (!evaluationMessagesAlreadySeen.Add(message))
                     {
@@ -599,6 +631,8 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     critical.ProjectFile = Intern(criticalArgs.ProjectFile);
                     critical.Subcategory = Intern(criticalArgs.Subcategory);
 
+                    Construction.PopulateWithExtendedData(critical, args);
+
                     nodeToAdd = critical;
                 }
                 else if (parent is Task task && task is CppAnalyzer.CppTask)
@@ -612,11 +646,38 @@ namespace Microsoft.Build.Logging.StructuredLogger
                 }
                 else
                 {
-                    nodeToAdd = new Message
+                    Message messageNode = null;
+                    string text = message;
+
+                    if (args is BuildMessageEventArgs buildMessageEventArgs)
                     {
-                        Text = message,
-                        IsLowRelevance = lowRelevance
-                    };
+                        if (!string.IsNullOrEmpty(buildMessageEventArgs.Code))
+                        {
+                            text = $"{buildMessageEventArgs.Code}: {text}";
+                            Intern(text);
+                        }
+
+                        if (!string.IsNullOrEmpty(buildMessageEventArgs.File))
+                        {
+                            messageNode = new MessageWithLocation
+                            {
+                                FilePath = buildMessageEventArgs.File,
+                                Line = buildMessageEventArgs.LineNumber
+                            };
+                        }
+                    }
+
+                    if (messageNode == null)
+                    {
+                        messageNode = new Message();
+                    }
+
+                    messageNode.Text = text;
+                    messageNode.IsLowRelevance = lowRelevance;
+
+                    Construction.PopulateWithExtendedData(messageNode, args);
+
+                    nodeToAdd = messageNode;
                 }
             }
 
@@ -629,9 +690,11 @@ namespace Microsoft.Build.Logging.StructuredLogger
             Folder results = task.Results;
             node = results ?? inputs;
 
+            int start;
+
             if (message.StartsWith("    ", StringComparison.Ordinal))
             {
-                message = message.Substring(4);
+                start = 4;
 
                 var parameter = node?.FindLastChild<Parameter>();
                 if (parameter != null)
@@ -639,6 +702,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
                     bool thereWasAConflict = Strings.IsThereWasAConflictPrefix(parameter.Name);
                     if (thereWasAConflict)
                     {
+                        message = message.Substring(start);
                         if (construction.Build.IsMSBuildVersionAtLeast(16, 9))
                         {
                             // https://github.com/KirillOsenkov/MSBuildStructuredLog/issues/443
@@ -652,48 +716,43 @@ namespace Microsoft.Build.Logging.StructuredLogger
                         return true;
                     }
 
-                    if (!string.IsNullOrWhiteSpace(message))
+                    node = parameter;
+
+                    if (message.StartsWith("        ", StringComparison.Ordinal))
                     {
-                        node = parameter;
+                        start = 8;
 
-                        if (message.StartsWith("    ", StringComparison.Ordinal))
+                        var lastItem = parameter.FindLastChild<Item>();
+
+                        // only indent if it's not a "For SearchPath..." message - that one needs to be directly under parameter
+                        // also don't indent if it's under AssemblyFoldersEx in Results
+                        if (lastItem != null &&
+                            !Strings.ForSearchPathPrefix.IsMatch(message, 8) &&
+                            !parameter.Name.StartsWith("AssemblyFoldersEx", StringComparison.Ordinal))
                         {
-                            message = message.Substring(4);
-
-                            var lastItem = parameter.FindLastChild<Item>();
-
-                            // only indent if it's not a "For SearchPath..." message - that one needs to be directly under parameter
-                            // also don't indent if it's under AssemblyFoldersEx in Results
-                            if (lastItem != null &&
-                                !Strings.ForSearchPathPrefix.IsMatch(message) &&
-                                !parameter.Name.StartsWith("AssemblyFoldersEx", StringComparison.Ordinal))
-                            {
-                                node = lastItem;
-                            }
+                            node = lastItem;
                         }
+                    }
 
-                        if (!string.IsNullOrEmpty(message))
+                    var equals = message.IndexOf('=');
+                    if (equals != -1 && message.IndexOfFirstLineBreak() == -1)
+                    {
+                        var kvp = TextUtilities.ParseNameValue(message, start);
+                        var metadata = new Metadata
                         {
-                            var equals = message.IndexOf('=');
-                            if (equals != -1 && message.IndexOfFirstLineBreak() == -1)
-                            {
-                                var kvp = TextUtilities.ParseNameValue(message);
-                                var metadata = new Metadata
-                                {
-                                    Name = Intern(kvp.Key.TrimEnd(space)),
-                                    Value = Intern(kvp.Value.TrimStart(space))
-                                };
-                                node.Children.Add(metadata);
-                                metadata.Parent = node;
-                            }
-                            else
-                            {
-                                node.AddChild(new Item
-                                {
-                                    Text = Intern(message)
-                                });
-                            }
-                        }
+                            Name = Intern(kvp.Key.TrimEnd(space)),
+                            Value = Intern(kvp.Value.TrimStart(space))
+                        };
+                        node.Children.Add(metadata);
+                        metadata.Parent = node;
+                    }
+                    else
+                    {
+                        message = message.Substring(start);
+                        node.AddChild(new Item
+                        {
+                            Text = Intern(message)
+                        });
                     }
 
                     return true;
@@ -876,6 +935,14 @@ namespace Microsoft.Build.Logging.StructuredLogger
             else if (message.StartsWith("Merging in runtimes", StringComparison.Ordinal))
             {
                 node = CreateFolder(node, "Merging in runtimes");
+            }
+            else if (message.StartsWith("Package source mapping", StringComparison.Ordinal))
+            {
+                node = CreateFolder(node, "Package source mapping");
+            }
+            else if (message.StartsWith("Restored ", StringComparison.Ordinal))
+            {
+                node = CreateFolder(node, "Restored");
             }
             else if (
                 message.StartsWith(Strings.RestoreTask_CheckingCompatibilityFor, StringComparison.Ordinal) ||
