@@ -8,7 +8,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
         private Target? target;
         public override string TypeName => nameof(EntryTarget);
         private Project? Project => project ??= GetNearestParent<Project>();
-        public Target? Target => target ??= Project?.FindFirstDescendant<Target>(t => t.Name == Name);
+        public Target? Target => target ??= Project?.FindFirstChild<Target>(t => t.Name == Name);
         public bool IsLowRelevance => Target?.IsLowRelevance ?? true;
         public string? DurationText => Target?.DurationText;
     }
