@@ -3,15 +3,19 @@
     /// <summary>
     /// Class representation of an item/property with associated metadata (if any).
     /// </summary>
-    public class Item : TextNode
+    public class Item : NamedNode
     {
-        public string NameAndEquals => string.IsNullOrWhiteSpace(Name) ? "" : Name + " = ";
+        public Item()
+        {
+            DisableChildrenCache = true;
+        }
 
         public override string TypeName => nameof(Item);
 
-        public override string ToString()
+        public string Text
         {
-            return string.IsNullOrWhiteSpace(Name) ? Text : NameAndEquals + Text;
+            get => Name;
+            set => Name = value;
         }
     }
 }
