@@ -1,6 +1,6 @@
 ﻿namespace Microsoft.Build.Logging.StructuredLogger
 {
-    public class NameValueNode : BaseNode, IHasTitle
+    public class NameValueNode : BaseNode
     {
         public string Name { get; set; }
         public string Value { get; set; }
@@ -9,10 +9,10 @@
         public bool IsValueShortened => Value != ShortenedValue;
 
         public override string TypeName => nameof(NameValueNode);
-        public override string ToString() => Name + " = " + Value;
-        public bool IsVisible { get { return true; } set { } }
-        public bool IsExpanded { get { return true; } set { } }
+        public override string Title => Name;
 
-        string IHasTitle.Title => Name;
+        public override string ToString() => Name + " = " + Value;
+        public bool IsVisible { get => true; set { } }
+        public bool IsExpanded { get => true; set { } }
     }
 }
