@@ -10,6 +10,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
         Expanded = 1 << 1,
         SearchResult = 1 << 2,
         ContainsSearchResult = 1 << 3,
-        LowRelevance = 1 << 4
+        LowRelevance = 1 << 4,
+        DisableChildrenCache = 1 << 5
     }
 }
