@@ -1,11 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Runtime.Versioning;
-
-namespace Microsoft.Build.Logging.StructuredLogger
+﻿namespace Microsoft.Build.Logging.StructuredLogger
 {
     public class Task : TimedNode, IHasSourceFile, IHasLineNumber
     {
@@ -13,31 +6,15 @@ namespace Microsoft.Build.Logging.StructuredLogger
         public string CommandLineArguments { get; set; }
         public string SourceFilePath { get; set; }
 
-        private string title;
-        public string Title
-        {
-            get
-            {
-                if (title == null)
-                {
-                    title = Name;
-                }
-
-                return title;
-            }
-
-            set
-            {
-                title = value;
-            }
-        }
-
         public override string TypeName => nameof(Task);
 
         public virtual bool IsDerivedTask => this.GetType() != typeof(Task);
 
         public int? LineNumber { get; set; }
+    }
 
-        public override string ToString() => Title;
+    public class MSBuildTask : Task
+    {
+        //public override string TypeName => nameof(MSBuildTask);
     }
 }
