@@ -2,6 +2,6 @@
 {
     public class CscTask : ManagedCompilerTask
     {
-        public override string TypeName => nameof(CscTask);
+        //public override string TypeName => nameof(CscTask);
     }
 }
