@@ -2,7 +2,7 @@
 {
     public class ManagedCompilerTask : Task
     {
-        public override string TypeName => nameof(ManagedCompilerTask);
+        //public override string TypeName => nameof(ManagedCompilerTask);
 
         private CompilationWrites? compilationWrites;
         public CompilationWrites? CompilationWrites
