@@ -1,23 +1,18 @@
 ﻿namespace Microsoft.Build.Logging.StructuredLogger
 {
-    public class TextNode : NamedNode
+    public class TextNode : TreeNode
     {
         public string Text { get; set; }
         public string ShortenedText => TextUtilities.ShortenValue(Text);
-        public bool IsTextShortened => Text != ShortenedText;
+        public bool IsTextShortened => Text != null && Text.Length != TextUtilities.GetShortenLength(Text);
 
-        protected override string GetTitle() => base.GetTitle() ?? Text;
+
         public override string TypeName => nameof(TextNode);
+        public override string Title => Text ?? TypeName;
 
         public override string ToString()
         {
-            var baseText = base.ToString();
-            if (string.IsNullOrEmpty(baseText))
-            {
-                return Text;
-            }
-
-            return baseText + " " + Text;
+            return Title;
         }
     }
 }
