@@ -1,9 +1,9 @@
-using System;
+﻿using System;
 using System.Threading;
 
 namespace Microsoft.Build.Logging.StructuredLogger
 {
-    public class Progress : IProgress<ProgressUpdate>, IProgress<double>
+    public class Progress : IProgress<ProgressUpdate>
     {
         public virtual CancellationToken CancellationToken { get; set; } = CancellationToken.None;
 
@@ -23,5 +23,6 @@ namespace Microsoft.Build.Logging.StructuredLogger
     public struct ProgressUpdate
     {
         public double Ratio;
+        public int BufferLength;
     }
 }
